Remove trace dependency to MlflowClient
