+++ mlflow/entities/span.py
@@ -404,24 +404,51 @@ class LiveSpan(Span):
         """
         self._span.add_event(event.name, event.attributes, event.timestamp)
 
-    def end(self, end_time: Optional[int] = None):
+    def end(self,
+        outputs: Optional[Any] = None,
+        attributes: Optional[dict[str, Any]] = None,
+        status: Optional[Union[SpanStatus, str]] = None,
+        end_time_ns: Optional[int] = None
+    ):
         """
-        End the span. This is a thin wrapper around the OpenTelemetry's end method but just
-        to handle the status update.
+        End the span.
 
-        This method should not be called directly by the user, only by called via fluent APIs
-        context exit or by MlflowClient APIs.
+        outputs: Outputs to set on the span.
+        attributes: A dictionary of attributes to set on the span. If the span already has
+            attributes, the new attributes will be merged with the existing ones. If the same
+            key already exists, the new value will overwrite the old one.
+        status: The status of the span. This can be a
+            :py:class:`SpanStatus <mlflow.entities.SpanStatus>` object or a string
+            representing the status code defined in
+            :py:class:`SpanStatusCode <mlflow.entities.SpanStatusCode>`
+            e.g. ``"OK"``, ``"ERROR"``. The default status is OK.
+        end_time_ns: The end time of the span in nano seconds since the UNIX epoch.
+            If not provided, the current time will be used.
 
         :meta private:
         """
-        # NB: In OpenTelemetry, status code remains UNSET if not explicitly set
-        # by the user. However, there is not way to set the status when using
-        # @mlflow.trace decorator. Therefore, we just automatically set the status
-        # to OK if it is not ERROR.
-        if self.status.status_code != SpanStatusCode.ERROR:
-            self.set_status(SpanStatus(SpanStatusCode.OK))
+        try:
+            self.set_attributes(attributes or {})
+            if outputs is not None:
+                self.set_outputs(outputs)
+            if status is not None:
+                self.set_status(status)
 
-        self._span.end(end_time=end_time)
+            # NB: In OpenTelemetry, status code remains UNSET if not explicitly set
+            # by the user. However, there is not way to set the status when using
+            # @mlflow.trace decorator. Therefore, we just automatically set the status
+            # to OK if it is not ERROR.
+            if self.status.status_code != SpanStatusCode.ERROR:
+                self.set_status(SpanStatus(SpanStatusCode.OK))
+
+            self._span.end(end_time=end_time_ns)
+
+        except Exception as e:
+            _logger.warning(
+                f"Failed to end span {self.span_id}: {e}. "
+                "For full traceback, set logging level to debug.",
+                exc_info=_logger.isEnabledFor(logging.DEBUG),
+            )
 
     def from_dict(cls, data: dict[str, Any]) -> "Span":
         raise NotImplementedError("The `from_dict` method is not supported for the LiveSpan class.")
@@ -473,7 +500,7 @@ class LiveSpan(Span):
         parent_span = trace_manager.get_span_from_id(request_id, parent_span_id)
 
         # Create a new span with the same name, parent, and start time
-        otel_span = mlflow.tracing.provider.start_detached_span(
+        otel_span = mlflow.tracing.provider._start_detached_otel_span(
             name=span.name,
             parent=parent_span._span if parent_span else None,
             start_time_ns=span.start_time_ns,
@@ -590,7 +617,12 @@ class NoOpSpan(Span):
     def add_event(self, event: SpanEvent):
         pass
 
-    def end(self):
+    def end(self,
+        outputs: Optional[Any] = None,
+        attributes: Optional[dict[str, Any]] = None,
+        status: Optional[Union[SpanStatus, str]] = None,
+        end_time_ns: Optional[int] = None
+    ):
         pass
