+++ mlflow/tracing/fluent.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import contextlib
-import functools
 import importlib
 import inspect
 import json
@@ -10,39 +9,28 @@ from contextvars import ContextVar
 from typing import TYPE_CHECKING, Any, Callable, Generator, Literal, Optional, Union
 
 from cachetools import TTLCache
+from mlflow.tracing.client import TracingClient
 from opentelemetry import trace as trace_api
 
-from mlflow import MlflowClient
 from mlflow.entities import NoOpSpan, SpanType, Trace
-from mlflow.entities.span import LiveSpan, create_mlflow_span
-from mlflow.entities.span_event import SpanEvent
-from mlflow.entities.span_status import SpanStatusCode
+from mlflow.entities.span import NO_OP_SPAN_REQUEST_ID, LiveSpan, create_mlflow_span
 from mlflow.entities.trace_status import TraceStatus
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import BAD_REQUEST
 from mlflow.store.tracking import SEARCH_TRACES_DEFAULT_MAX_RESULTS
 from mlflow.tracing import provider
 from mlflow.tracing.constant import (
-    STREAM_CHUNK_EVENT_NAME_FORMAT,
-    STREAM_CHUNK_EVENT_VALUE_KEY,
     SpanAttributeKey,
 )
-from mlflow.tracing.provider import (
-    is_tracing_enabled,
-    safe_set_span_in_context,
-)
+from mlflow.tracing.provider import is_tracing_enabled
 from mlflow.tracing.trace_manager import InMemoryTraceManager
 from mlflow.tracing.utils import (
     SPANS_COLUMN_NAME,
-    TraceJSONEncoder,
-    capture_function_input_args,
     encode_span_id,
-    end_client_span_or_trace,
+    exclude_immutable_tags,
     get_otel_attribute,
-    start_client_span_or_trace,
 )
 from mlflow.tracing.utils.search import extract_span_inputs_outputs, traces_to_df
-from mlflow.tracking.fluent import _get_experiment_id
 from mlflow.utils import get_results_from_paginated_fn
 from mlflow.utils.annotations import deprecated, experimental
 
@@ -160,207 +148,19 @@ def trace(
             single value to be set as the span output.
         model_id: If specified, associates the span with the given model ID.
     """
+    from mlflow.tracing.utils._wrapper import wrap_function, wrap_generator
 
     def decorator(fn):
         if inspect.isgeneratorfunction(fn) or inspect.isasyncgenfunction(fn):
-            return _wrap_generator(fn, name, span_type, attributes, output_reducer)
+            return wrap_generator(fn, name, span_type, attributes, output_reducer)
         else:
             if output_reducer is not None:
                 raise MlflowException.invalid_parameter_value(
                     "The output_reducer argument is only supported for generator functions."
                 )
-            return _wrap_function(fn, name, span_type, attributes, model_id)
+            return wrap_function(fn, name, span_type, attributes, model_id)
 
     return decorator(func) if func else decorator
-
-
-def _wrap_function(
-    fn: Callable,
-    name: Optional[str] = None,
-    span_type: str = SpanType.UNKNOWN,
-    attributes: Optional[dict[str, Any]] = None,
-    model_id: Optional[str] = None,
-) -> Callable:
-    class _WrappingContext:
-        # define the wrapping logic as a coroutine to avoid code duplication
-        # between sync and async cases
-        @staticmethod
-        def _wrapping_logic(fn, args, kwargs):
-            span_name = name or fn.__name__
-
-            with start_span(
-                name=span_name, span_type=span_type, attributes=attributes, model_id=model_id
-            ) as span:
-                span.set_attribute(SpanAttributeKey.FUNCTION_NAME, fn.__name__)
-                span.set_inputs(capture_function_input_args(fn, args, kwargs))
-                result = yield  # sync/async function output to be sent here
-                span.set_outputs(result)
-                try:
-                    yield result
-                except GeneratorExit:
-                    # Swallow `GeneratorExit` raised when the generator is closed
-                    pass
-
-        def __init__(self, fn, args, kwargs):
-            self.coro = self._wrapping_logic(fn, args, kwargs)
-
-        def __enter__(self):
-            next(self.coro)
-            return self.coro
-
-        def __exit__(self, exc_type, exc_value, traceback):
-            # Since the function call occurs outside the coroutine,
-            # if an exception occurs, we need to throw it back in, so that
-            # we return control to the coro (in particular, so that the __exit__'s
-            # of start_span and OTel's use_span can execute).
-            if exc_type is not None:
-                self.coro.throw(exc_type, exc_value, traceback)
-            self.coro.close()
-
-    if inspect.iscoroutinefunction(fn):
-
-        async def wrapper(*args, **kwargs):
-            with _WrappingContext(fn, args, kwargs) as wrapping_coro:
-                return wrapping_coro.send(await fn(*args, **kwargs))
-    else:
-
-        def wrapper(*args, **kwargs):
-            with _WrappingContext(fn, args, kwargs) as wrapping_coro:
-                return wrapping_coro.send(fn(*args, **kwargs))
-
-    return functools.wraps(fn)(wrapper)
-
-
-def _wrap_generator(
-    fn: Callable,
-    name: Optional[str] = None,
-    span_type: str = SpanType.UNKNOWN,
-    attributes: Optional[dict[str, Any]] = None,
-    output_reducer: Optional[Callable] = None,
-) -> Callable:
-    """
-    Wrap a generator function to create a span.
-
-    Generator functions need special handling because of its lazy evaluation nature.
-    Let's say we have a generator function like this:
-
-    ```
-    @mlflow.trace
-    def generate_stream():
-        # B
-        for i in range(10):
-            # C
-            yield i * 2
-        # E
-
-
-    stream = generate_stream()
-    # A
-    for chunk in stream:
-        # D
-        pass
-    # F
-    ```
-
-    The execution order is A -> B -> C -> D -> C -> D -> ... -> E -> F.
-    The span should only be "active" at B, C, and E, namely, when the code execution
-    is inside the generator function. Otherwise it will create wrong span tree, or
-    even worse, leak span context and pollute subsequent traces.
-    """
-
-    def _start_stream_span(fn, args, kwargs):
-        try:
-            return start_client_span_or_trace(
-                client=MlflowClient(),
-                name=name or fn.__name__,
-                parent_span=get_current_active_span(),
-                span_type=span_type,
-                attributes=attributes,
-                inputs=capture_function_input_args(fn, args, kwargs),
-            )
-        except Exception as e:
-            _logger.debug(f"Failed to start stream span: {e}")
-            return NoOpSpan()
-
-    def _end_stream_span(
-        span: LiveSpan,
-        outputs: Optional[list[Any]] = None,
-        output_reducer: Optional[Callable] = None,
-        error: Optional[Exception] = None,
-    ):
-        client = MlflowClient()
-        if error:
-            span.add_event(SpanEvent.from_exception(error))
-            end_client_span_or_trace(client, span, status=SpanStatusCode.ERROR)
-            return
-
-        if output_reducer:
-            try:
-                outputs = output_reducer(outputs)
-            except Exception as e:
-                _logger.debug(f"Failed to reduce outputs from stream: {e}")
-        end_client_span_or_trace(client, span, outputs=outputs)
-
-    def _record_chunk_event(span: LiveSpan, chunk: Any, chunk_index: int):
-        try:
-            event = SpanEvent(
-                name=STREAM_CHUNK_EVENT_NAME_FORMAT.format(index=chunk_index),
-                # OpenTelemetry SpanEvent only support str-str key-value pairs for attributes
-                attributes={STREAM_CHUNK_EVENT_VALUE_KEY: json.dumps(chunk, cls=TraceJSONEncoder)},
-            )
-            span.add_event(event)
-        except Exception as e:
-            _logger.debug(f"Failing to record chunk event for span {span.name}: {e}")
-
-    if inspect.isgeneratorfunction(fn):
-
-        def wrapper(*args, **kwargs):
-            span = _start_stream_span(fn, args, kwargs)
-            generator = fn(*args, **kwargs)
-
-            i = 0
-            outputs = []
-            while True:
-                try:
-                    # NB: Set the span to active only when the generator is running
-                    with safe_set_span_in_context(span):
-                        value = next(generator)
-                except StopIteration:
-                    break
-                except Exception as e:
-                    _end_stream_span(span, error=e)
-                    raise e
-                else:
-                    outputs.append(value)
-                    _record_chunk_event(span, value, i)
-                    yield value
-                    i += 1
-            _end_stream_span(span, outputs, output_reducer)
-    else:
-
-        async def wrapper(*args, **kwargs):
-            span = _start_stream_span(fn, args, kwargs)
-            generator = fn(*args, **kwargs)
-
-            i = 0
-            outputs = []
-            while True:
-                try:
-                    with safe_set_span_in_context(span):
-                        value = await generator.__anext__()
-                except StopAsyncIteration:
-                    break
-                except Exception as e:
-                    _end_stream_span(span, error=e)
-                    raise e
-                else:
-                    outputs.append(value)
-                    _record_chunk_event(span, value, i)
-                    yield value
-                    i += 1
-            _end_stream_span(span, outputs, output_reducer)
-
-    return functools.wraps(fn)(wrapper)
 
 
 @contextlib.contextmanager
@@ -456,6 +256,66 @@ def start_span(
             _logger.debug(f"Failed to end span {mlflow_span.span_id}.", exc_info=True)
 
 
+def start_span_no_context(
+    name: str,
+    span_type: str = SpanType.UNKNOWN,
+    parent_span: Optional[LiveSpan] = None,
+    inputs: Optional[Any] = None,
+    attributes: Optional[dict[str, str]] = None,
+    tags: Optional[dict[str, str]] = None,
+    experiment_id: Optional[str] = None,
+    start_time_ns: Optional[int] = None,
+) -> LiveSpan:
+    """
+    TBA
+    """
+    # If parent span is no-op span, the child should also be no-op too
+    if parent_span and parent_span.request_id == NO_OP_SPAN_REQUEST_ID:
+        return NoOpSpan()
+
+    try:
+        # Create new trace and a root span
+        # Once OTel span is created, SpanProcessor.on_start is invoked
+        # TraceInfo is created and logged into backend store inside on_start method
+        otel_span = provider.start_detached_span(
+            name, experiment_id=experiment_id, start_time_ns=start_time_ns
+        )
+
+        if parent_span:
+            request_id = parent_span.request_id
+        else:
+            request_id = get_otel_attribute(otel_span, SpanAttributeKey.REQUEST_ID)
+
+        mlflow_span = create_mlflow_span(otel_span, request_id, span_type)
+
+        # # If the span is a no-op span i.e. tracing is disabled, do nothing
+        if isinstance(mlflow_span, NoOpSpan):
+            return mlflow_span
+
+        if inputs is not None:
+            mlflow_span.set_inputs(inputs)
+        mlflow_span.set_attributes(attributes or {})
+
+        trace_manager = InMemoryTraceManager.get_instance()
+        if tags := exclude_immutable_tags(tags or {}):
+            # Update trace tags for trace in in-memory trace manager
+            with trace_manager.get_trace(request_id) as trace:
+                trace.info.tags.update(tags)
+
+        # Register new span in the in-memory trace manager
+        trace_manager.register_span(mlflow_span)
+
+        return mlflow_span
+    except Exception as e:
+        _logger.warning(
+            f"Failed to start span {name}: {e}. "
+            "For full traceback, set logging level to debug.",
+            exc_info=_logger.isEnabledFor(logging.DEBUG),
+        )
+    return NoOpSpan()
+
+
+
 def get_trace(request_id: str) -> Optional[Trace]:
     """
     Get a trace by the given request ID if it exists.
@@ -488,7 +348,7 @@ def get_trace(request_id: str) -> Optional[Trace]:
     request_id = _EVAL_REQUEST_ID_TO_TRACE_ID.get(request_id) or request_id
 
     try:
-        return MlflowClient().get_trace(request_id, display=False)
+        return TracingClient().get_trace(request_id)
     except MlflowException as e:
         _logger.warning(
             f"Failed to get trace from the tracking store: {e}"
@@ -622,6 +482,8 @@ def search_traces(
         mlflow.search_traces(run_id=run.info.run_id, return_type="list")
 
     """
+    from mlflow.tracking.fluent import _get_experiment_id
+
     if return_type not in ["pandas", "list"]:
         raise MlflowException.invalid_parameter_value(
             f"Invalid return type: {return_type}. Return type must be either 'pandas' or 'list'."
@@ -650,7 +512,7 @@ def search_traces(
             )
 
     def pagination_wrapper_func(number_to_get, next_page_token):
-        return MlflowClient().search_traces(
+        return TracingClient().search_traces(
             experiment_ids=experiment_ids,
             run_id=run_id,
             max_results=number_to_get,
@@ -765,7 +627,7 @@ def get_last_active_trace() -> Optional[Trace]:
     """
     if _LAST_ACTIVE_TRACE_ID_GLOBAL is not None:
         try:
-            return MlflowClient().get_trace(_LAST_ACTIVE_TRACE_ID_GLOBAL, display=False)
+            return TracingClient().get_trace(_LAST_ACTIVE_TRACE_ID_GLOBAL)
         except:
             _logger.debug(
                 "Failed to get the last active trace with "
@@ -981,9 +843,8 @@ def add_trace(trace: Union[Trace, dict[str, Any]], target: Optional[LiveSpan] = 
         # If there is no target span, create a new root span named "Remote Trace <...>"
         # and put the remote trace under it. This design aims to keep the trace export
         # logic simpler and consistent, rather than directly exporting the remote trace.
-        client = MlflowClient()
         remote_root_span = trace.data.spans[0]
-        span = client.start_trace(
+        span = start_span_no_context(
             name=f"Remote Trace <{remote_root_span.name}>",
             inputs=remote_root_span.inputs,
             attributes={
@@ -999,7 +860,7 @@ def add_trace(trace: Union[Trace, dict[str, Any]], target: Optional[LiveSpan] = 
             target_request_id=span.request_id,
             target_parent_span_id=span.span_id,
         )
-        client.end_trace(
+        span.end(
             request_id=span.request_id,
             status=trace.info.status,
             outputs=remote_root_span.outputs,
@@ -1061,21 +922,20 @@ def log_trace(
 
         print(trace.data.intermediate_outputs)
     """
-    client = MlflowClient()
     if intermediate_outputs:
         if attributes:
             attributes.update(SpanAttributeKey.INTERMEDIATE_OUTPUTS, intermediate_outputs)
         else:
             attributes = {SpanAttributeKey.INTERMEDIATE_OUTPUTS: intermediate_outputs}
 
-    span = client.start_trace(
+    span = start_span_no_context(
         name=name,
         inputs=request,
         attributes=attributes,
         tags=tags,
         start_time_ns=start_time_ms * 1000000 if start_time_ms else None,
     )
-    client.end_trace(
+    span.end(
         request_id=span.request_id,
         outputs=response,
         end_time_ns=(start_time_ms + execution_time_ms) * 1000000
