+++ mlflow/tracking/client.py
@@ -78,13 +78,11 @@ from mlflow.store.model_registry import (
     SEARCH_REGISTERED_MODEL_MAX_RESULTS_DEFAULT,
 )
 from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT, SEARCH_TRACES_DEFAULT_MAX_RESULTS
-from mlflow.tracing.constant import (
-    TRACE_REQUEST_ID_PREFIX,
-    SpanAttributeKey,
-)
+from mlflow.tracing.fluent import start_span_no_context
+from mlflow.tracing.client import TracingClient
+from mlflow.tracing.constant import TRACE_REQUEST_ID_PREFIX
 from mlflow.tracing.display import get_display_handler
 from mlflow.tracing.trace_manager import InMemoryTraceManager
-from mlflow.tracing.utils import exclude_immutable_tags, get_otel_attribute
 from mlflow.tracking._model_registry import DEFAULT_AWAIT_MAX_SLEEP_SECONDS
 from mlflow.tracking._model_registry import utils as registry_utils
 from mlflow.tracking._model_registry.client import ModelRegistryClient
@@ -160,6 +158,8 @@ class MlflowClient:
         final_tracking_uri = utils._resolve_tracking_uri(tracking_uri)
         self._registry_uri = registry_utils._resolve_registry_uri(registry_uri, tracking_uri)
         self._tracking_client = TrackingServiceClient(final_tracking_uri)
+        self._tracing_client = TracingClient(tracking_uri)
+
         # `MlflowClient` also references a `ModelRegistryClient` instance that is provided by the
         # `MlflowClient._get_registry_client()` method. This `ModelRegistryClient` is not explicitly
         # defined as an instance variable in the `MlflowClient` constructor; an instance variable
@@ -781,7 +781,7 @@ class MlflowClient:
     ##### Tracing #####
 
     def _upload_trace_data(self, trace_info: TraceInfo, trace_data: TraceData) -> None:
-        return self._tracking_client._upload_trace_data(trace_info, trace_data)
+        return self._tracing_client._upload_trace_data(trace_info, trace_data)
 
     def delete_traces(
         self,
@@ -833,7 +833,7 @@ class MlflowClient:
             # Delete traces based on request_ids
             client.delete_traces(experiment_id="0", request_ids=["id_1", "id_2"])
         """
-        return self._tracking_client.delete_traces(
+        return self._tracing_client.delete_traces(
             experiment_id=experiment_id,
             max_timestamp_millis=max_timestamp_millis,
             max_traces=max_traces,
@@ -866,7 +866,7 @@ class MlflowClient:
                 "the search_traces() API."
             )
 
-        trace = self._tracking_client.get_trace(request_id)
+        trace = self._tracing_client.get_trace(request_id)
         if display:
             get_display_handler().display_traces([trace])
         return trace
@@ -912,23 +912,7 @@ class MlflowClient:
             some store implementations may not support pagination and thus the returned token would
             not be meaningful in such cases.
         """
-        if model_id is not None:
-            if filter_string:
-                raise MlflowException(
-                    message=(
-                        "Cannot specify both `model_id` or `filter_string` in the search_traces "
-                        "call."
-                    ),
-                    error_code=INVALID_PARAMETER_VALUE,
-                )
-
-            filter_string = (
-                f"request_metadata.`mlflow.modelId` = '{model_id}'"
-                if sql_warehouse_id is None
-                else None
-            )
-
-        return self._tracking_client.search_traces(
+        return self._tracing_client.search_traces(
             experiment_ids=experiment_ids,
             filter_string=filter_string,
             max_results=max_results,
@@ -1021,40 +1005,15 @@ class MlflowClient:
                 error_code=BAD_REQUEST,
             )
 
-        try:
-            # Create new trace and a root span
-            # Once OTel span is created, SpanProcessor.on_start is invoked
-            # TraceInfo is created and logged into backend store inside on_start method
-            otel_span = mlflow.tracing.provider.start_detached_span(
-                name, experiment_id=experiment_id, start_time_ns=start_time_ns
-            )
-            request_id = get_otel_attribute(otel_span, SpanAttributeKey.REQUEST_ID)
-            mlflow_span = create_mlflow_span(otel_span, request_id, span_type)
-
-            # # If the span is a no-op span i.e. tracing is disabled, do nothing
-            if isinstance(mlflow_span, NoOpSpan):
-                return mlflow_span
-
-            if inputs is not None:
-                mlflow_span.set_inputs(inputs)
-            mlflow_span.set_attributes(attributes or {})
-
-            trace_manager = InMemoryTraceManager.get_instance()
-            tags = exclude_immutable_tags(tags or {})
-            # Update trace tags for trace in in-memory trace manager
-            with trace_manager.get_trace(request_id) as trace:
-                trace.info.tags.update(tags)
-            # Register new span in the in-memory trace manager
-            trace_manager.register_span(mlflow_span)
-
-            return mlflow_span
-        except Exception as e:
-            _logger.warning(
-                f"Failed to start trace {name}: {e}. "
-                "For full traceback, set logging level to debug.",
-                exc_info=_logger.isEnabledFor(logging.DEBUG),
-            )
-            return NoOpSpan()
+        return start_span_no_context(
+            name=name,
+            span_type=span_type,
+            inputs=inputs,
+            attributes=attributes,
+            tags=tags,
+            experiment_id=experiment_id,
+            start_time_ns=start_time_ns,
+        )
 
     def end_trace(
         self,
@@ -1107,7 +1066,8 @@ class MlflowClient:
                     error_code=INVALID_PARAMETER_VALUE,
                 )
 
-        self.end_span(request_id, root_span_id, outputs, attributes, status, end_time_ns)
+        root_span = trace_manager.get_span_from_id(request_id, root_span_id)
+        root_span.end(outputs, attributes, status, end_time_ns)
 
     @experimental
     def _log_trace(self, trace: Trace) -> str:
@@ -1133,13 +1093,13 @@ class MlflowClient:
         # Create trace info entry in the backend
         # Note that the backend generates a new request ID for the trace. Currently there is
         # no way to insert the trace with a specific request ID given by the user.
-        new_info = self._tracking_client.start_trace(
+        new_info = self._tracing_client.start_trace(
             experiment_id=experiment_id,
             timestamp_ms=trace.info.timestamp_ms,
             request_metadata={},
             tags={},
         )
-        self._tracking_client.end_trace(
+        self._tracing_client.end_trace(
             request_id=new_info.request_id,
             # Compute the end time of the original trace
             timestamp_ms=trace.info.timestamp_ms + trace.info.execution_time_ms,
@@ -1294,26 +1254,13 @@ class MlflowClient:
                 error_code=RESOURCE_DOES_NOT_EXIST,
             )
 
-        try:
-            otel_span = mlflow.tracing.provider.start_detached_span(
-                name=name,
-                parent=parent_span._span,
-                start_time_ns=start_time_ns,
-            )
-
-            span = create_mlflow_span(otel_span, request_id, span_type)
-            span.set_attributes(attributes or {})
-            if inputs is not None:
-                span.set_inputs(inputs)
-            trace_manager.register_span(span)
-            return span
-        except Exception as e:
-            _logger.warning(
-                f"Failed to start span {name}: {e}. "
-                "For full traceback, set logging level to debug.",
-                exc_info=_logger.isEnabledFor(logging.DEBUG),
-            )
-            return NoOpSpan()
+        return start_span_no_context(
+            name=name,
+            span_type=span_type,
+            parent_span=parent_span,
+            inputs=inputs,
+            attributes=attributes,
+        )
 
     def end_span(
         self,
@@ -1342,78 +1289,14 @@ class MlflowClient:
             end_time_ns: The end time of the span in nano seconds since the UNIX epoch.
                 If not provided, the current time will be used.
         """
-        if request_id == NO_OP_SPAN_REQUEST_ID:
-            return
-
-        trace_manager = InMemoryTraceManager.get_instance()
-        span = trace_manager.get_span_from_id(request_id, span_id)
-
-        if span is None:
-            raise MlflowException(
-                f"Span with ID {span_id} is not found or already finished.",
-                error_code=RESOURCE_DOES_NOT_EXIST,
-            )
-        span.set_attributes(attributes or {})
-        if outputs is not None:
-            span.set_outputs(outputs)
-        span.set_status(status)
-
-        try:
-            span.end(end_time=end_time_ns)
-        except Exception as e:
-            _logger.warning(
-                f"Failed to end span {span_id}: {e}. "
-                "For full traceback, set logging level to debug.",
-                exc_info=_logger.isEnabledFor(logging.DEBUG),
-            )
-
-    def _start_tracked_trace(
-        self,
-        experiment_id: str,
-        timestamp_ms: int,
-        request_metadata: Optional[dict[str, str]] = None,
-        tags: Optional[dict[str, str]] = None,
-    ) -> TraceInfo:
-        """
-        Start an initial TraceInfo object in the backend store.
-
-        Args:
-            experiment_id: String id of the experiment for this run.
-            timestamp_ms: Start time of the trace, in milliseconds since the UNIX epoch.
-            request_metadata: Metadata of the trace.
-            tags: Tags of the trace.
-
-        Returns:
-            The created TraceInfo object.
-        """
-        # Some tags like mlflow.runName are immutable once logged in tracking server.
-        return self._tracking_client.start_trace(
-            experiment_id=experiment_id,
-            timestamp_ms=timestamp_ms,
-            request_metadata=request_metadata or {},
-            tags=tags or {},
+        span = InMemoryTraceManager.get_instance().get_span_from_id(request_id, span_id)
+        span.end(
+            outputs=outputs,
+            attributes=attributes,
+            status=status,
+            end_time_ns=end_time_ns,
         )
 
-    def _upload_ended_trace_info(
-        self,
-        trace_info: TraceInfo,
-    ) -> TraceInfo:
-        """
-        Update the TraceInfo object in the backend store with the completed trace info.
-
-        Args:
-            trace_info: Updated TraceInfo object to be stored in the backend store.
-
-        Returns:
-            The updated TraceInfo object.
-        """
-        return self._tracking_client.end_trace(
-            request_id=trace_info.request_id,
-            timestamp_ms=trace_info.timestamp_ms + trace_info.execution_time_ms,
-            status=trace_info.status,
-            request_metadata=trace_info.request_metadata,
-            tags=trace_info.tags or {},
-        )
 
     def set_trace_tag(self, request_id: str, key: str, value: str):
         """
@@ -1441,27 +1324,7 @@ class MlflowClient:
             value: The string value of the tag. Must be at most 250 characters long, otherwise
                 it will be truncated when stored.
         """
-        if key.startswith("mlflow."):
-            raise MlflowException(
-                f"Tags starting with 'mlflow.' are reserved and cannot be set. "
-                f"Attempted to set tag with key '{key}' on trace with ID '{request_id}'.",
-                error_code=INVALID_PARAMETER_VALUE,
-            )
-
-        if not isinstance(value, str):
-            _logger.warning(
-                "Received non-string value for trace tag. Please note that non-string tag values"
-                "will automatically be stringified when the trace is logged."
-            )
-
-        # Trying to set the tag on the active trace first
-        with InMemoryTraceManager.get_instance().get_trace(request_id) as trace:
-            if trace:
-                trace.info.tags[key] = str(value)
-                return
-
-        # If the trace is not active, try to set the tag on the trace in the backend
-        self._tracking_client.set_trace_tag(request_id, key, value)
+        self._tracing_client.set_trace_tag(request_id, key, value)
 
     def delete_trace_tag(self, request_id: str, key: str) -> None:
         """
@@ -1487,20 +1350,7 @@ class MlflowClient:
             key: The string key of the tag. Must be at most 250 characters long, otherwise
                 it will be truncated when stored.
         """
-        # Trying to delete the tag on the active trace first
-        with InMemoryTraceManager.get_instance().get_trace(request_id) as trace:
-            if trace:
-                if key in trace.info.tags:
-                    trace.info.tags.pop(key)
-                    return
-                else:
-                    raise MlflowException(
-                        f"Tag with key {key} not found in trace with ID {request_id}.",
-                        error_code=RESOURCE_DOES_NOT_EXIST,
-                    )
-
-        # If the trace is not active, try to delete the tag on the trace in the backend
-        self._tracking_client.delete_trace_tag(request_id, key)
+        self._tracing_client.delete_trace_tag(request_id, key)
 
     def log_assessment(
         self,
@@ -1513,8 +1363,7 @@ class MlflowClient:
         metadata: Optional[dict[str, Any]] = None,
         span_id: Optional[str] = None,
     ) -> Assessment:
-        assessment = Assessment(
-            # assessment_id must be None when creating a new assessment
+        return self._tracing_client.log_assessment(
             trace_id=trace_id,
             name=name,
             source=source,
@@ -1524,7 +1373,6 @@ class MlflowClient:
             metadata=metadata,
             span_id=span_id,
         )
-        return self._tracking_client.create_assessment(assessment)
 
     def update_assessment(
         self,
@@ -1536,7 +1384,7 @@ class MlflowClient:
         rationale: Optional[str] = None,
         metadata: Optional[dict[str, Any]] = None,
     ) -> Assessment:
-        return self._tracking_client.update_assessment(
+        return self._tracing_client.update_assessment(
             trace_id=trace_id,
             assessment_id=assessment_id,
             name=name,
@@ -1547,7 +1395,7 @@ class MlflowClient:
         )
 
     def delete_assessment(self, trace_id: str, assessment_id: str) -> None:
-        return self._tracking_client.delete_assessment(trace_id, assessment_id)
+        return self._tracing_client.delete_assessment(trace_id, assessment_id)
 
     def search_experiments(
         self,
